Move exception defintions to domain
