+++ backend/src/main/kotlin/com/tngtech/apicenter/backend/connector/rest/controller/VersionController.kt
@@ -2,7 +2,7 @@ package com.tngtech.apicenter.backend.connector.rest.controller
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.dataformat.yaml.YAMLMapper
-import com.tngtech.apicenter.backend.config.SpecificationNotFoundException
+import com.tngtech.apicenter.backend.domain.exceptions.SpecificationNotFoundException
 
 import com.tngtech.apicenter.backend.domain.entity.Version
 import com.tngtech.apicenter.backend.connector.rest.dto.VersionDto
